Homework 10 - add gitignore file, move POJO to dto folder
