+++ src/main/java/dto/CityCatalog.java
@@ -1,4 +1,4 @@
-package homework_10_Api;
+package dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
